Move nuxt app directory contents outside
