+++ apps/website/components/MascotSVG.vue
