+++ apps/website/components/VMascot.vue
