--- apps/website/shared/constants/images.ts
@@ -1,2 +1,0 @@
-/** The file extension used for website images */
-export const imageFileExtension = "avif";
