--- apps/website/shared/types/editor.ts
@@ -1,4 +1,0 @@
-export type {
-  JSONRecipeIngredientAttributes,
-  JSONRecipeIngredientContent,
-} from "@wizard/content";
