--- apps/website/shared/types/image.ts
@@ -1,10 +1,0 @@
-export type ImagePurpose = "cover" | "preview" | "instruction";
-export type AspectRatio = "square" | "portrait";
-
-export type ImageVariant = {
-  src: string;
-  srcSet: string;
-  sizes: string;
-};
-
-export type ImageVariants = Record<ImagePurpose, Record<AspectRatio, ImageVariant>>;
