--- apps/website/shared/types/version.ts
@@ -1,4 +1,0 @@
-export type Version = {
-  build: string;
-  searchIndex: string;
-};
