--- apps/website/shared/utils/error.ts
@@ -1,3 +1,0 @@
-export const throwExpression = (errorMessage: string): never => {
-  throw new Error(errorMessage);
-};
