+++ apps/website/types/version.ts
@@ -1,0 +1,6 @@
+export type AppVersion = {
+  /** The identifier of the current build. */
+  build: string;
+  /** The current version of the search index. */
+  searchIndex: string;
+};
